Add files via upload
Here are the code files for the classes so far. This is the general idea I've come up with for the moment, I am still planning to add some more complimentary functions and such but this is should provide a basic start at least. If you have any ideas on things to add or change, please let me know.