Add files via upload
I have edited the Product and Sizes class a little. I am going to add a Shoe class; the Shoe object will be the one added to the cart instead of the Product object as the Shoe object will be instantiated with only the information necessary for checkout (Name, color, size, upc, and price).